Conflicts:
	Server/FoodBack/src/main/java/com/lpro/fbrest/api/Meal.java
	Server/FoodBack/src/main/java/com/lpro/fbrest/api/Orders_meal.java
	Server/FoodBack/src/main/java/com/lpro/fbrest/db/MealDAO.java
	Server/FoodBack/src/main/java/com/lpro/fbrest/resources/MealsResource.java
	Server/FoodBack/src/main/java/com/lpro/fbrest/service/MealService.java

+++ Server/FoodBack/src/main/java/com/lpro/fbrest/db/MealDAO.java
@@ -1,5 +1,6 @@
 package com.lpro.fbrest.db;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.skife.jdbi.v2.sqlobject.Bind;
@@ -10,26 +11,28 @@ import org.skife.jdbi.v2.sqlobject.customizers.RegisterMapper;
 
 import com.lpro.fbrest.api.Meal;
 
+/**
+ * Mapper for Meal Class
+ */
 @RegisterMapper(MealMapper.class)
 public interface MealDAO {
 	
 	/**
-	 * @param meal
-	 * @param price
-	 * @param establishment_id
-	 * @return
+	 * @param meal Meal name
+	 * @param price Meal price
+	 * @param establishment_id ID of establishment
+	 * @return ID of inserted meal
 	 */
 	@SqlUpdate("INSERT INTO meal "
 			+ "VALUES (DEFAULT, :meal, :price, :establishment_id)")
 	@GetGeneratedKeys
 	long insertMeal(@Bind("meal") String meal,
-					@Bind("price") int price,
+					@Bind("price") BigDecimal price,
 					@Bind("establishment_id") long establishment_id);
-
-
+	
 	/**
-	 * @param establishment_id
-	 * @return
+	 * @param establishment_id ID of establishment
+	 * @return List of meals from the specified establishment
 	 */
 	@SqlQuery("SELECT id, meal, price, establishment_id "
 			+ "FROM meal "
@@ -37,15 +40,13 @@ public interface MealDAO {
 	List<Meal> getAllMeals(@Bind("establishment_id") long establishment_id);
 
 	/**
-	 * @param establishment_id
-	 * @param meal
-	 * @return
+	 * @param establishment_id ID of establishment
+	 * @param meal Meal name
+	 * @return Meal with specified name
 	 */
 	@SqlQuery("SELECT id, meal, price, establishment_id "
 			+ "FROM meal "
 			+ "WHERE establishment_id = :establishment_id AND meal = :meal")
 	Meal getMealbyName(@Bind("establishment_id") long establishment_id, @Bind("meal") String meal);
-
-	
 
 }
